+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.1.1-SNAPSHOT
 * Make it possible to use the plugin with Java 7
+* Provide default name patterns for classes to be excluded from report generation.
+Currently, these patterns include *Android*, *Dagger 2* and *Butter Knife* generated classes.
 
 ## Version 0.1.0 *(2015-09-13)*
 * Initial release